+++ sdk/monitor/azure-monitor-opentelemetry/azure/monitor/opentelemetry/_configure.py
@@ -151,7 +151,22 @@ def _setup_tracing(configurations: Dict[str, ConfigurationValue]):
     tracer_provider.add_span_processor(bsp)
     set_tracer_provider(tracer_provider)
     if _is_instrumentation_enabled(configurations, _AZURE_SDK_INSTRUMENTATION_NAME):
-        settings.tracing_implementation = OpenTelemetrySpan
+        try:
+            from azure.core.settings import settings
+            from azure.core.tracing.ext.opentelemetry_span import OpenTelemetrySpan
+            settings.tracing_implementation = OpenTelemetrySpan
+        except ImportError as ex:
+            # This could possibly be due to breaking change in upstream OpenTelemetry
+            # Advise user to upgrade to latest OpenTelemetry version
+            _logger.warning(  # pylint: disable=do-not-log-exceptions-if-not-debug
+                "Exception occured when importing Azure SDK Tracing. Please upgrade to the latest OpenTelemetry version: %s.",
+                ex,
+            )
+        except Exception as ex:  # pylint: disable=broad-except
+            _logger.warning(  # pylint: disable=do-not-log-exceptions-if-not-debug
+                "Exception occured when setting Azure SDK Tracing: %s.",
+                ex,
+            )
 
 
 def _setup_logging(configurations: Dict[str, ConfigurationValue]):