+++ sdk/monitor/azure-monitor-opentelemetry/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Other Changes
 
-- Defensively import and vendor experimental components from OpenTelemetry
+- Defensively import and vendor experimental components from OpenTelemetry and azure.core + upgrade dependencies to latest OpenTelemetry
     ([#40866](https://github.com/Azure/azure-sdk-for-python/pull/40866))
 
 ## 1.6.9 (2025-05-12)